requirements.txt
torch==2.7.0
torchvision==0.22.0
pandas
numpy
matplotlib
Pillow
scikit-learn
